+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-03T08:15:58.819Z",
+    "2026-03-03T10:04:52.010Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {