+++ packages/orq-rc/src/models/operations/fileget.ts
@@ -94,7 +94,7 @@ export const FileGetResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-03-03T08:16:02.263Z",
+    "2026-03-03T10:04:55.606Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {