+++ packages/orq-rc/RELEASES.md
@@ -7419,3 +7419,13 @@ Based on:
 - [typescript v4.5.0-rc.22] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.22 - packages/orq-rc
+
+## 2026-03-03 10:07:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.739.0 (2.846.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.23] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.23 - packages/orq-rc